+++ src/Object/ClusterCredentials.php
@@ -33,6 +33,7 @@ use Teknoo\East\Foundation\Normalizer\Object\NormalizableInterface;
 use Teknoo\East\Common\Contracts\Object\IdentifiedObjectInterface;
 use Teknoo\East\Common\Contracts\Object\TimestampableInterface;
 use Teknoo\East\Common\Object\ObjectTrait;
+use Teknoo\East\Paas\Object\Traits\ExportConfigurationsTrait;
 use Teknoo\Immutable\ImmutableTrait;
 use Teknoo\East\Paas\Contracts\Object\IdentityInterface;
 
@@ -54,6 +55,7 @@ class ClusterCredentials implements
     use ObjectTrait;
     use ImmutableTrait;
     use GroupsTrait;
+    use ExportConfigurationsTrait;
 
     private string $caCertificate = '';