+++ src/Object/SshIdentity.php
@@ -32,6 +32,7 @@ use Teknoo\East\Foundation\Normalizer\Object\NormalizableInterface;
 use Teknoo\East\Common\Contracts\Object\IdentifiedObjectInterface;
 use Teknoo\East\Common\Object\ObjectTrait;
 use Teknoo\East\Common\Contracts\Object\TimestampableInterface;
+use Teknoo\East\Paas\Object\Traits\ExportConfigurationsTrait;
 use Teknoo\Immutable\ImmutableTrait;
 use Teknoo\East\Paas\Contracts\Object\IdentityInterface;
 
@@ -53,6 +54,7 @@ class SshIdentity implements
     use ObjectTrait;
     use ImmutableTrait;
     use GroupsTrait;
+    use ExportConfigurationsTrait;
 
     private string $name = '';