+++ tests/src/Object/XRegistryAuthTest.php
@@ -26,6 +26,7 @@ declare(strict_types=1);
 namespace Teknoo\Tests\East\Paas\Object;
 
 use PHPUnit\Framework\TestCase;
+use ReflectionClass;
 use Teknoo\East\Foundation\Normalizer\EastNormalizerInterface;
 use Teknoo\East\Paas\Object\XRegistryAuth;
 use Teknoo\Tests\East\Common\Object\Traits\ObjectTestTrait;
@@ -35,6 +36,7 @@ use Teknoo\Tests\East\Common\Object\Traits\ObjectTestTrait;
  * @license     http://teknoo.software/license/mit         MIT License
  * @author      Richard Déloge <richard@teknoo.software>
  * @covers \Teknoo\East\Paas\Object\XRegistryAuth
+ * @covers \Teknoo\East\Paas\Object\Traits\ExportConfigurationsTrait
  */
 class XRegistryAuthTest extends TestCase
 {
@@ -170,6 +172,17 @@ class XRegistryAuthTest extends TestCase
                 $normalizer,
                 ['groups' => 'api']
             )
+        );
+    }
+
+    public function testSetExportConfiguration()
+    {
+        XRegistryAuth::setExportConfiguration($conf = ['name' => ['all']]);
+        $rc = new ReflectionClass(XRegistryAuth::class);
+
+        self::assertEquals(
+            $conf,
+            $rc->getStaticPropertyValue('exportConfigurations'),
         );
     }
 }